+++ lib/src/ui/widget/visual/decorated/decorated_app.widget.dart
@@ -3,9 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 
 class DecoratedApp<B extends RootBLoC> extends StatelessWidget {
-  DecoratedApp({
+  const DecoratedApp({
     super.key,
-    B? rootBLoC,
+    required B rootBLoC,
     @Deprecated('已无作用, 直接使用DecoratedApp的参数即可') this.app = NIL,
     @Deprecated('暂无作用') this.preventTextScale = false,
     this.onGenerateTitle,
@@ -40,7 +40,7 @@ class DecoratedApp<B extends RootBLoC> extends StatelessWidget {
     this.showPerformanceOverlay = false,
     this.showSemanticsDebugger = false,
     this.restorationScopeId,
-  }) : rootBLoC = rootBLoC ?? get();
+  }) : rootBLoC = rootBLoC;
 
   final B rootBLoC;
   final bool preventTextScale;